scripts/migrate.sh: install & run migrate bot

+++ scripts/migrate.sh
@@ -5,6 +5,9 @@
 #
 # [!] Existing database will be dropped, so use with caution
 #
+# Make sure to upgrade your existing Hitchwiki installation to MediaWiki 1.28
+# before generating the database dumps for this script
+#
 
 if [ ! -f Vagrantfile ]; then # an arbirtrary file that appears only once in the whole repository tree
     echo "ERROR: Bad working directory ($(pwd))."
@@ -16,10 +19,23 @@ fi
 source "scripts/_path_resolve.sh"
 source "$SCRIPTDIR/_settings.sh"
 
+echo "Remove existing Hitchwiki images folder..."
+rm -rf "$WIKIDIR/images"
+echo
+
+echo "Extract old English Hitchwiki images tarball..."
+DUMPFILE="$ROOTDIR/dumps/old-images.tar.gz"
+if [ ! -f "$DUMPFILE" ]; then
+    echo "File $DUMPFILE not found"
+    exit 1
+fi
+tar -xzf "$DUMPFILE" -C "$WIKIDIR"
+echo
+
 echo "Drop $HW__db__database database and recreate it..."
 mysql -u$HW__db__username -p$HW__db__password -e "DROP DATABASE IF EXISTS $HW__db__database"
 mysql -u$HW__db__username -p$HW__db__password -e "CREATE DATABASE $HW__db__database CHARACTER SET utf8 COLLATE utf8_general_ci"
-echo ""
+echo
 
 echo "Import old English Hitchwiki SQL dump..."
 DUMPFILE="$ROOTDIR/dumps/old-hitchwiki_en.sql"
@@ -28,12 +44,12 @@ if [ ! -f "$DUMPFILE" ]; then
     exit 1
 fi
 cat "$DUMPFILE" | mysql -u$HW__db__username -p$HW__db__password $HW__db__database
-echo ""
+echo
 
 echo "Drop hitchwiki_maps database and recreate it..."
 mysql -u$HW__db__username -p$HW__db__password -e "DROP DATABASE IF EXISTS hitchwiki_maps"
 mysql -u$HW__db__username -p$HW__db__password -e "CREATE DATABASE hitchwiki_maps CHARACTER SET utf8 COLLATE utf8_general_ci"
-echo ""
+echo
 
 echo "Import old Hitchwiki Maps SQL dump..."
 DUMPFILE="$ROOTDIR/dumps/old-hitchwiki_maps.sql"
@@ -42,12 +58,12 @@ if [ ! -f "$DUMPFILE" ]; then
     exit 1
 fi
 cat "$DUMPFILE" | mysql -u$HW__db__username -p$HW__db__password hitchwiki_maps
-echo ""
+echo
 
 echo "Drop hitchwiki_rate database and recreate it..."
 mysql -u$HW__db__username -p$HW__db__password -e "DROP DATABASE IF EXISTS hitchwiki_rate"
 mysql -u$HW__db__username -p$HW__db__password -e "CREATE DATABASE hitchwiki_rate CHARACTER SET utf8 COLLATE utf8_general_ci"
-echo ""
+echo
 
 echo "Import old Hitchwiki Rate SQL dump..."
 DUMPFILE="$ROOTDIR/dumps/old-hitchwiki_rate.sql"
@@ -56,18 +72,18 @@ if [ ! -f "$DUMPFILE" ]; then
     exit 1
 fi
 cat "$DUMPFILE" | mysql -u$HW__db__username -p$HW__db__password hitchwiki_rate
-echo ""
+echo
 
 echo "Update MediaWiki..."
 cd "$WIKIDIR"
 php maintenance/update.php --quick --conf "$MWCONFFILE"
-echo ""
+echo
 
 # Pre-populate the antispoof (MW extension) table with your wiki's existing usernames
 echo "Pre-populate antispoof table..."
 cd "$WIKIDIR"
 php extensions/AntiSpoof/maintenance/batchAntiSpoof.php
-echo ""
+echo
 
 # Import Semantic pages
 cd "$ROOTDIR"
@@ -75,8 +91,44 @@ bash "$SCRIPTDIR/import_pages.sh"
 
 echo "Import interwiki table..."
 mysql -u$HW__db__username -p$HW__db__password $HW__db__database < "$SCRIPTDIR/configs/interwiki.sql"
-echo ""
+echo
 
-#Create basic users
+# Create basic users
 cd "$ROOTDIR"
 bash "$SCRIPTDIR/create_users.sh"
+
+# Install dependencies of migrate bot
+echo "Install dependencies of migrate bot..."
+bash "$SCRIPTDIR/bot/bot_install.sh"
+echo
+
+# Run article migrate bot
+echo "Run article migrate bot: annotate place articles with geographical Semantic MW templates (this might take a while)..."
+cd "$SCRIPTDIR/bot"
+python pywikibot-core/pwb.py migrate/articlemigrate.py
+echo
+
+# Run spot migrate bot
+echo "Run spot migrate bot: turn spots from the old hitchwiki_maps DB into Semantic MW articles (this might take a while)..."
+cd "$SCRIPTDIR/bot"
+python pywikibot-core/pwb.py migrate/articlemigrate.py
+echo
+
+# Run extra migrate bot
+echo "Run extra migrate bot: copy comments, ratings and waiting times from the old DB tables into the new DB..."
+cd "$SCRIPTDIR/bot"
+python pywikibot-core/pwb.py migrate/extramigrate.py
+echo
+
+# Drop hitchwiki_maps DB
+echo "Drop no longer needed hitchwiki_maps database..."
+mysql -u$HW__db__username -p$HW__db__password -e "DROP DATABASE IF EXISTS hitchwiki_maps"
+echo
+
+# Drop hitchwiki_rate DB
+echo "Drop no longer needed hitchwiki_rate database..."
+mysql -u$HW__db__username -p$HW__db__password -e "DROP DATABASE IF EXISTS hitchwiki_rate"
+echo
+
+echo "Done! Grrrreat success."
+echo
